refs #466 fixed compilation issue
git-svn-id: https://devel.codership.com/svn/galera/branches/0.8@2152 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ galera/src/trx_handle.hpp
@@ -297,7 +297,7 @@ namespace galera
         bool                   certified_;
         bool                   committed_;
         long                   gcs_handle_;
-        long long              timestamp_;
+        int64_t                timestamp_;
         Mac                    mac_;
 
         //
